requirements.txt
python-telegram-bot==20.3
gspread==5.6.2
oauth2client==4.1.3
Flask==2.2.5
gunicorn==20.1.0


SQLAlchemy~=2.0.36